Fixed bug where map was emitting constant errors that it couldnt find the layer 'clusters'
